refactor: replace the_movie_db with reelix_manager module, prune unused GQL fields
- Rename the_movie_db/ to services/reelix_manager/ with new types.rs
- Add Clone to GqlVideoBlob, remove deprecated fields from MovieResponse, TvResponse, SeasonResponse, SeasonEpisode
- Update all consumers: commands (auth, general, rip), services (ftp, upload_recovery), state (job_state, title_video), templates (+21 more)
