+++ src-tauri/src/services/plex.rs
@@ -1,13 +1,13 @@
 use crate::services::reelix_manager::ReelixManager;
 use crate::state::AppState;
-use crate::the_movie_db;
+use crate::reelix_manager::{MovieResponse, SeasonResponse, TvResponse};
 use std::collections::HashSet;
 use tauri::{AppHandle, Manager};
 
 pub fn find_movie(
     app_handle: &AppHandle,
     id: u32,
-) -> Result<(the_movie_db::MovieResponse, bool), crate::services::reelix_manager::Error> {
+) -> Result<(MovieResponse, bool), crate::services::reelix_manager::Error> {
     let state: tauri::State<AppState> = app_handle.state::<AppState>();
     let manager = ReelixManager::new(&state);
     manager.find_movie(id)
@@ -16,7 +16,7 @@ pub fn find_movie(
 pub fn find_tv(
     app_handle: &AppHandle,
     id: u32,
-) -> Result<the_movie_db::TvResponse, crate::services::reelix_manager::Error> {
+) -> Result<TvResponse, crate::services::reelix_manager::Error> {
     let state: tauri::State<AppState> = app_handle.state::<AppState>();
     let manager = ReelixManager::new(&state);
     manager.find_tv(id)
@@ -26,7 +26,7 @@ pub fn find_season(
     app_handle: &AppHandle,
     tv_id: u32,
     season_number: u32,
-) -> Result<(the_movie_db::SeasonResponse, HashSet<u32>), crate::services::reelix_manager::Error> {
+) -> Result<(SeasonResponse, HashSet<u32>), crate::services::reelix_manager::Error> {
     let state: tauri::State<AppState> = app_handle.state::<AppState>();
     let manager = ReelixManager::new(&state);
     manager.find_season(tv_id, season_number)
