--- src-tauri/src/the_movie_db/mod.rs
@@ -1,3 +1,0 @@
-pub mod models;
-
-pub use models::*;
