--- src-tauri/templates/the_movie_db/index.html
@@ -1,3 +1,0 @@
-<div id="{{ self.dom_id() }}">
-  <p class="text-muted">Unable to connect to the server API. Please check your server connection.</p>
-</div>
